rename: rename LottoLank to LottoRank
